CLLM/requirements.txt
typer[all]>=0.9.0
rich>=13.0.0
google-genai>=1.14.0
python-dotenv>=0.21.0
prompt_toolkit
platformdirs